changed int to long

+++ Program.cs
@@ -7,9 +7,9 @@
             int a = 1;  
             int b = 2;
             int c = a * b;
-            int bigNumber1 = 7864188;
-            int bigNumber2 = 4851184;
-            int multBigNumbers = bigNumber1 * bigNumber2;
+            long bigNumber1 = 7864188;
+            long bigNumber2 = 4851184;
+            long multBigNumbers = bigNumber1 * bigNumber2;
             Console.WriteLine("Hello, World! " + a + " * " + b + " = " + c);
             Console.WriteLine("Hold on to your hats! Multiplying BIIIG NUMBERS!!!: " + multBigNumbers);
         }
